move error to a new folder
