+++ src/__test__/components/PlatformError.test.jsx
@@ -11,9 +11,9 @@ import { act } from 'react-dom/test-utils';
 
 import PlatformError from 'components/PlatformError';
 
-import WorkResponseError from 'utils/http/errors/WorkResponseError';
-import WorkGenericError from 'utils/http/errors/WorkGenericError';
-import WorkTimeoutError from 'utils/http/errors/WorkTimeoutError';
+import WorkResponseError from 'utils/errors/http//WorkResponseError';
+import WorkGenericError from 'utils/errors/http//WorkGenericError';
+import WorkTimeoutError from 'utils/errors/http//WorkTimeoutError';
 
 import '__test__/test-utils/setupTests';
