+++ src/__test__/components/plots/ContinuousEmbeddingPlot.test.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import { act } from 'react-dom/test-utils';
 import { screen, render } from '@testing-library/react';
-import userEvent from '@testing-library/user-event';
 import { Provider } from 'react-redux';
 import { makeStore } from 'redux/store';
 import fetchMock, { enableFetchMocks } from 'jest-fetch-mock';
@@ -18,7 +17,7 @@ import createTestComponentFactory from '__test__/test-utils/testComponentFactory
 import fake from '__test__/test-utils/constants';
 import mockEmbedding from '__test__/data/embedding.json';
 import mockGeneExpression from '__test__/data/gene_expression.json';
-import WorkResponseError from 'utils/http/errors/WorkResponseError';
+import WorkResponseError from 'utils/errors/http//WorkResponseError';
 
 enableFetchMocks();
