+++ src/utils/errors/upload/SampleValidationError.js
@@ -1,0 +1,9 @@
+class SampleValidationError extends Error {
+  constructor(error) {
+    super(error);
+
+    this.name = 'SampleValidationError';
+    this.error = error;
+  }
+}
+module.exports = SampleValidationError;
